Remove more duplicate initialization code
