+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/grid/GridScrollModeUtil.java
@@ -1,0 +1,27 @@
+package com.refinedmods.refinedstorage2.platform.common.internal.grid;
+
+import com.refinedmods.refinedstorage2.platform.api.grid.GridScrollMode;
+
+import net.minecraft.network.FriendlyByteBuf;
+
+public final class GridScrollModeUtil {
+    private GridScrollModeUtil() {
+    }
+
+    public static GridScrollMode getMode(byte mode) {
+        if (mode == 0) {
+            return GridScrollMode.GRID_TO_INVENTORY;
+        } else if (mode == 1) {
+            return GridScrollMode.GRID_TO_CURSOR;
+        }
+        return GridScrollMode.INVENTORY_TO_GRID;
+    }
+
+    public static void writeMode(FriendlyByteBuf buf, GridScrollMode mode) {
+        switch (mode) {
+            case GRID_TO_INVENTORY -> buf.writeByte(0);
+            case GRID_TO_CURSOR -> buf.writeByte(1);
+            case INVENTORY_TO_GRID -> buf.writeByte(2);
+        }
+    }
+}
