+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/packet/c2s/GridScrollPacket.java
@@ -3,6 +3,7 @@ package com.refinedmods.refinedstorage2.platform.fabric.packet.c2s;
 import com.refinedmods.refinedstorage2.platform.api.grid.GridScrollMode;
 import com.refinedmods.refinedstorage2.platform.api.grid.ItemGridEventHandler;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
+import com.refinedmods.refinedstorage2.platform.common.internal.grid.GridScrollModeUtil;
 import com.refinedmods.refinedstorage2.platform.common.util.PacketUtil;
 
 import net.fabricmc.fabric.api.networking.v1.PacketSender;
@@ -17,7 +18,7 @@ public class GridScrollPacket implements ServerPlayNetworking.PlayChannelHandler
     @Override
     public void receive(MinecraftServer server, ServerPlayer player, ServerGamePacketListenerImpl handler, FriendlyByteBuf buf, PacketSender responseSender) {
         ItemResource itemResource = PacketUtil.readItemResource(buf);
-        GridScrollMode mode = getMode(buf.readByte());
+        GridScrollMode mode = GridScrollModeUtil.getMode(buf.readByte());
         int slot = buf.readInt();
 
         server.execute(() -> {
@@ -26,22 +27,5 @@ public class GridScrollPacket implements ServerPlayNetworking.PlayChannelHandler
                 gridEventHandler.onScroll(itemResource, mode, slot);
             }
         });
-    }
-
-    private GridScrollMode getMode(byte mode) {
-        if (mode == 0) {
-            return GridScrollMode.GRID_TO_INVENTORY;
-        } else if (mode == 1) {
-            return GridScrollMode.GRID_TO_CURSOR;
-        }
-        return GridScrollMode.INVENTORY_TO_GRID;
-    }
-
-    public static void writeMode(FriendlyByteBuf buf, GridScrollMode mode) {
-        switch (mode) {
-            case GRID_TO_INVENTORY -> buf.writeByte(0);
-            case GRID_TO_CURSOR -> buf.writeByte(1);
-            case INVENTORY_TO_GRID -> buf.writeByte(2);
-        }
     }
 }
