+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/packet/c2s/GridScrollPacket.java
@@ -3,6 +3,7 @@ package com.refinedmods.refinedstorage2.platform.forge.packet.c2s;
 import com.refinedmods.refinedstorage2.platform.api.grid.GridScrollMode;
 import com.refinedmods.refinedstorage2.platform.api.grid.ItemGridEventHandler;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
+import com.refinedmods.refinedstorage2.platform.common.internal.grid.GridScrollModeUtil;
 import com.refinedmods.refinedstorage2.platform.common.util.PacketUtil;
 
 import java.util.function.Supplier;
@@ -27,14 +28,14 @@ public class GridScrollPacket {
     public static GridScrollPacket decode(FriendlyByteBuf buf) {
         return new GridScrollPacket(
                 PacketUtil.readItemResource(buf),
-                getMode(buf.readByte()),
+                GridScrollModeUtil.getMode(buf.readByte()),
                 buf.readInt()
         );
     }
 
     public static void encode(GridScrollPacket packet, FriendlyByteBuf buf) {
         PacketUtil.writeItemResource(buf, packet.itemResource);
-        writeMode(buf, packet.mode);
+        GridScrollModeUtil.writeMode(buf, packet.mode);
         buf.writeInt(packet.slot);
     }
 
@@ -50,23 +51,6 @@ public class GridScrollPacket {
         AbstractContainerMenu screenHandler = player.containerMenu;
         if (screenHandler instanceof ItemGridEventHandler gridEventHandler) {
             gridEventHandler.onScroll(packet.itemResource, packet.mode, packet.slot);
-        }
-    }
-
-    private static GridScrollMode getMode(byte mode) {
-        if (mode == 0) {
-            return GridScrollMode.GRID_TO_INVENTORY;
-        } else if (mode == 1) {
-            return GridScrollMode.GRID_TO_CURSOR;
-        }
-        return GridScrollMode.INVENTORY_TO_GRID;
-    }
-
-    public static void writeMode(FriendlyByteBuf buf, GridScrollMode mode) {
-        switch (mode) {
-            case GRID_TO_INVENTORY -> buf.writeByte(0);
-            case GRID_TO_CURSOR -> buf.writeByte(1);
-            case INVENTORY_TO_GRID -> buf.writeByte(2);
         }
     }
 }
